+++ src/main/java/com/braintreegateway/AndroidPayDetails.java
@@ -7,6 +7,7 @@ public class AndroidPayDetails {
     private String last4;
     private String sourceCardType;
     private String sourceCardLast4;
+    private String sourceDescription;
     private String virtualCardType;
     private String virtualCardLast4;
     private String expirationMonth;
@@ -19,6 +20,7 @@ public class AndroidPayDetails {
     public AndroidPayDetails(NodeWrapper node) {
         this.sourceCardType = node.findString("source-card-type");
         this.sourceCardLast4 = node.findString("source-card-last-4");
+        this.sourceDescription = node.findString("source-description");
         this.virtualCardType = node.findString("virtual-card-type");
         this.virtualCardLast4 = node.findString("virtual-card-last-4");
         this.cardType = this.virtualCardType;
@@ -45,6 +47,10 @@ public class AndroidPayDetails {
 
     public String getSourceCardLast4() {
         return sourceCardLast4;
+    }
+
+    public String getSourceDescription() {
+        return sourceDescription;
     }
 
     public String getVirtualCardType() {