+++ src/main/java/com/braintreegateway/ApplePayDetails.java
@@ -5,6 +5,7 @@ import com.braintreegateway.util.NodeWrapper;
 public class ApplePayDetails {
     private String cardType;
     private String paymentInstrumentName;
+    private String sourceDescription;
     private String cardholderName;
     private String expirationMonth;
     private String expirationYear;
@@ -14,6 +15,7 @@ public class ApplePayDetails {
     public ApplePayDetails(NodeWrapper node) {
         cardType = node.findString("card-type");
         paymentInstrumentName = node.findString("payment-instrument-name");
+        sourceDescription = node.findString("source-description");
         cardholderName = node.findString("cardholder-name");
         expirationMonth = node.findString("expiration-month");
         expirationYear = node.findString("expiration-year");
@@ -31,6 +33,10 @@ public class ApplePayDetails {
 
     public String getPaymentInstrumentName() {
         return paymentInstrumentName;
+    }
+
+    public String getSourceDescription() {
+        return sourceDescription;
     }
 
     public String getCardholderName() {