+++ src/test/java/com/braintreegateway/integrationtest/TransactionIT.java
@@ -3,6 +3,7 @@ package com.braintreegateway.integrationtest;
 import com.braintreegateway.*;
 import com.braintreegateway.SandboxValues.CreditCardNumber;
 import com.braintreegateway.SandboxValues.TransactionAmount;
+import com.braintreegateway.SandboxValues.AmexRewardsRequestId;
 import com.braintreegateway.TestingGateway;
 import com.braintreegateway.exceptions.ForgedQueryStringException;
 import com.braintreegateway.exceptions.NotFoundException;
@@ -671,6 +672,7 @@ public class TransactionIT implements MerchantAccountTestConstants {
         assertNotNull(transaction.getApplePayDetails());
         assertNotNull(transaction.getApplePayDetails().getCardType());
         assertNotNull(transaction.getApplePayDetails().getPaymentInstrumentName());
+        assertNotNull(transaction.getApplePayDetails().getSourceDescription());
         assertNotNull(transaction.getApplePayDetails().getExpirationMonth());
         assertNotNull(transaction.getApplePayDetails().getExpirationYear());
         assertNotNull(transaction.getApplePayDetails().getCardholderName());
@@ -678,8 +680,8 @@ public class TransactionIT implements MerchantAccountTestConstants {
     }
 
     @Test
-    public void saleWithAndroidPayCardNonce() {
-        String androidPayCardNonce = Nonce.AndroidPay;
+    public void saleWithAndroidPayProxyCardNonce() {
+        String androidPayCardNonce = Nonce.AndroidPayDiscover;
 
         TransactionRequest request = new TransactionRequest().
             amount(SandboxValues.TransactionAmount.AUTHORIZE.amount).
@@ -699,6 +701,39 @@ public class TransactionIT implements MerchantAccountTestConstants {
         assertNotNull(androidPayDetails.getImageUrl());
         assertNotNull(androidPayDetails.getSourceCardType());
         assertNotNull(androidPayDetails.getSourceCardLast4());
+        assertNotNull(androidPayDetails.getSourceDescription());
+        assertNotNull(androidPayDetails.getVirtualCardType());
+        assertNotNull(androidPayDetails.getVirtualCardLast4());
+        assertNotNull(androidPayDetails.getGoogleTransactionId());
+        assertNotNull(androidPayDetails.getCardType());
+        assertNotNull(androidPayDetails.getLast4());
+        assertNotNull(androidPayDetails.getExpirationMonth());
+        assertNotNull(androidPayDetails.getExpirationYear());
+    }
+
+    @Test
+    public void saleWithAndroidPayNetworkTokenNonce() {
+        String androidPayCardNonce = Nonce.AndroidPayMasterCard;
+
+        TransactionRequest request = new TransactionRequest().
+            amount(SandboxValues.TransactionAmount.AUTHORIZE.amount).
+            paymentMethodNonce(androidPayCardNonce);
+
+        Result<Transaction> result = gateway.transaction().sale(request);
+        assertTrue(result.isSuccess());
+        Transaction transaction = result.getTarget();
+
+        assertEquals(PaymentInstrumentType.ANDROID_PAY_CARD, transaction.getPaymentInstrumentType());
+
+        assertNotNull(transaction.getAndroidPayDetails());
+        AndroidPayDetails androidPayDetails = transaction.getAndroidPayDetails();
+
+        assertNull(androidPayDetails.getToken());
+        assertNotNull(androidPayDetails.getBin());
+        assertNotNull(androidPayDetails.getImageUrl());
+        assertNotNull(androidPayDetails.getSourceCardType());
+        assertNotNull(androidPayDetails.getSourceCardLast4());
+        assertNotNull(androidPayDetails.getSourceDescription());
         assertNotNull(androidPayDetails.getVirtualCardType());
         assertNotNull(androidPayDetails.getVirtualCardLast4());
         assertNotNull(androidPayDetails.getGoogleTransactionId());
@@ -795,6 +830,118 @@ public class TransactionIT implements MerchantAccountTestConstants {
         assertFalse(result.isSuccess());
         assertEquals(ValidationErrorCode.TRANSACTION_THREE_D_SECURE_TRANSACTION_DATA_DOESNT_MATCH_VERIFY,
                 result.getErrors().forObject("transaction").onField("threeDSecureToken").get(0).getCode());
+    }
+
+    @Test
+    public void saleWithAmexRewards() {
+        TransactionRequest request = new TransactionRequest().
+            amount(TransactionAmount.AUTHORIZE.amount).
+            creditCard().
+                number(CreditCardNumber.AMEX.number).
+                expirationDate("12/2020").
+                done()
+            .options().
+                submitForSettlement(true).
+                amexRewards().
+                    requestId("ABC123").
+                    points("1000").
+                    currencyAmount("10.00").
+                    currencyIsoCode("USD").
+                    done().
+                done();
+
+        Result<Transaction> result = gateway.transaction().sale(request);
+        assertTrue(result.isSuccess());
+        Transaction transaction = result.getTarget();
+
+        assertEquals(Transaction.Status.SUBMITTED_FOR_SETTLEMENT, transaction.getStatus());
+        assertEquals("success", transaction.getAmexRewardsResponse());
+    }
+
+    @Test
+    public void saleErrorWithIneligibleAmexRewards() {
+        TransactionRequest request = new TransactionRequest().
+            amount(TransactionAmount.AUTHORIZE.amount).
+            creditCard().
+                number(CreditCardNumber.AMEX.number).
+                expirationDate("12/2020").
+                done()
+            .options().
+                submitForSettlement(true).
+                amexRewards().
+                    requestId(AmexRewardsRequestId.CARD_INELIGIBLE.requestId).
+                    points("1000").
+                    currencyAmount("10.00").
+                    currencyIsoCode("USD").
+                    done().
+                done();
+
+        Result<Transaction> result = gateway.transaction().sale(request);
+        assertTrue(result.isSuccess());
+        Transaction transaction = result.getTarget();
+
+        assertEquals(Transaction.Status.SUBMITTED_FOR_SETTLEMENT, transaction.getStatus());
+        assertEquals("RDM2002 Card is not eligible for redemption", transaction.getAmexRewardsResponse());
+    }
+
+    @Test
+    public void submitForSettlementWithAmexRewards() {
+        String nonce = TestHelper.generateOneTimePayPalNonce(gateway);
+
+        TransactionRequest request = new TransactionRequest().
+            amount(new BigDecimal("100.00")).
+            creditCard().
+                number(CreditCardNumber.AMEX.number).
+                expirationDate("12/2020").
+                done()
+            .options().
+                amexRewards().
+                    requestId("ABC123").
+                    points("1000").
+                    currencyAmount("10.00").
+                    currencyIsoCode("USD").
+                    done().
+                done();
+
+        Result<Transaction> saleResult = gateway.transaction().sale(request);
+
+        assertTrue(saleResult.isSuccess());
+        assertEquals(Transaction.Status.AUTHORIZED, saleResult.getTarget().getStatus());
+
+        Result<Transaction> submitForSettlementResult = gateway.transaction().submitForSettlement(saleResult.getTarget().getId());
+        assertTrue(submitForSettlementResult.isSuccess());
+        assertEquals(Transaction.Status.SUBMITTED_FOR_SETTLEMENT, submitForSettlementResult.getTarget().getStatus());
+        assertEquals("success", submitForSettlementResult.getTarget().getAmexRewardsResponse());
+    }
+
+    @Test
+    public void submitForSettlementErrorWithIneligibleAmexRewards() {
+        String nonce = TestHelper.generateOneTimePayPalNonce(gateway);
+
+        TransactionRequest request = new TransactionRequest().
+            amount(new BigDecimal("100.00")).
+            creditCard().
+                number(CreditCardNumber.AMEX.number).
+                expirationDate("12/2020").
+                done()
+            .options().
+                amexRewards().
+                    requestId(AmexRewardsRequestId.CARD_INELIGIBLE.requestId).
+                    points("1000").
+                    currencyAmount("10.00").
+                    currencyIsoCode("USD").
+                    done().
+                done();
+
+        Result<Transaction> saleResult = gateway.transaction().sale(request);
+
+        assertTrue(saleResult.isSuccess());
+        assertEquals(Transaction.Status.AUTHORIZED, saleResult.getTarget().getStatus());
+
+        Result<Transaction> submitForSettlementResult = gateway.transaction().submitForSettlement(saleResult.getTarget().getId());
+        assertTrue(submitForSettlementResult.isSuccess());
+        assertEquals(Transaction.Status.SUBMITTED_FOR_SETTLEMENT, submitForSettlementResult.getTarget().getStatus());
+        assertEquals("RDM2002 Card is not eligible for redemption", submitForSettlementResult.getTarget().getAmexRewardsResponse());
     }
 
     @Test